User account db work
Manually brought over all the stuff I did without the refactoring so as not to break anything before our presentation. I also added the user display page to the nav bar so it is accessible.
Note that in this commit, I had to bump our java version up from 1.7 to 1.8 so that we can use lambas.