Implemented sign-in split for new/existing users
This was extremely annoying to figure out. I am honestly writing this for my future self to remember what happened here.
Guides highly recommend to use LiveData when getting data from the db. This is because it allows you to see your data update in real time and deals with some concurrency/threading stuff.
You cannot do db operations on the main thread. This is why you see a bunch of weird databaseWriteExecutor calls in Repository classes. That allows db stuff to occur on a seperate thread. However, what I wanted to do was not working with that approach. I wanted to check the db to see if a user existed in it and then do different things based on the result. This doesnt work with our current framework.

So I had to bring in two new classes: Callable and Future. Cant say I completely understand what they do but they replace the databaseWriteExecutor stuff so I would assume they are using other threads and blocking the main thread until the data is ready. By doing this, I was finally able to implement the SQL query and successfully branch what the app does based on if a user is new or existing!