Added SonarCloud Property -- Project Version
I decided to add versioning based on our Sprints. So 0.0.1 would be the very first sprint of our project. Sprint 2 would be 0.0.2 and so on.
Since we are on Sprint 4, we are on 0.0.4 and I will increment once we start the next sprint. 
I will use the middle digit for more grand advancements, such as moving into the Spring semester. 
If we were to ever go public and release, we could use the first digit. 
This is just something I created for SonarCloud and can be changed as needed.